Initial checkin
It's a file splitter tool which chunks the large csv files into smaller csv files so that it can be processed efficiently in memory by any programming language or techniques. i.e there are data files of hug size which needs to be processed to dump data in database (i.e. bulk upload, SSIS etc)